protect source folder, and name some guaranteed targets

+++ BashScripts/copycogs.sh
@@ -3,14 +3,36 @@
 # tidy!
 clear
 
-# where our libraries are located
-libPath=~/'Documents/Development/WoW Projects/AzeriteUI/AzeriteUI/back-end'
+# Name of the source project
+# This will never be deleted. Again. /facepalm
+sourceProject='AzeriteUI'
+
+# Name of known target projects
+# I need this when I accidentally delete 
+# the libraries without adding them back in. 
+declare -A targetProjects=( 
+	["GoldpawUI"]="true" 
+	["SimpleClassPower"]="true"
+)
 
 # where our source addon files are located
-wowProjects=~/'Documents/Development/WoWProjects'
+projectPath=~/'Documents/Development/WoWProjects'
+
+# where our libraries are located
+libPath="$projectPath/$sourceProject/$sourceProject/back-end"
+
+if [ ! -d "$libPath" ]
+then
+	echo "Source libraries are missing!"
+	echo "Path: $libPath"
+	echo " "
+	read -rp "Press Enter to exit!"
+	clear
+	exit 0
+fi 
 
 # iterate through the project folder
-for dir in "$wowProjects"/*/								
+for dir in "$projectPath"/*/								
 do
 	
 	# full path to the addon project directory
@@ -23,19 +45,19 @@ do
 		projectPath=${subDir%*/}
 
 		# name of the addon project directory
-		addonDir=${projectPath##*/}
+		projectName=${projectPath##*/}
 
 		# path to this addon's library folder
 		backEnd="$projectPath/back-end"
 
 		#echo "$backEnd"
 
-		if [ -d "$backEnd" ]
-		then 
+		if [ -d "$backEnd" ] || [ "${targetProjects["$projectName"]}" == "true" ]
+		then
 
-			if [ "$backEnd" != "$libPath" ]
+			if [ "$backEnd" != "$libPath" ] && [ "$projectName" != "$sourceProject" ]
 			then
-				echo "...Updating CogWheel libraries in '$addonDir'" 
+				echo "...Updating CogWheel libraries in '$projectName'" 
 
 				echo "......Clearing out the old"
 				rm -rf "$backEnd"
@@ -50,4 +72,5 @@ do
 done 
 
 read -rp "Press Enter to exit!"
+clear
 exit 0
